Flytter konvertør over i ny fil
