+++ src/main/kotlin/no/nav/k9/pleiepengersyktbarn/soknad/PleiepengerSyktBarnRoutes.kt
@@ -3,29 +3,20 @@ package no.nav.k9.pleiepengersyktbarn.soknad
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.convertValue
 import kotlinx.coroutines.reactive.awaitFirst
+import no.nav.k9.*
+import no.nav.k9.mappe.*
+import no.nav.k9.søknad.JsonUtils
+import no.nav.k9.søknad.ValideringsFeil
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
+import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import org.springframework.web.reactive.function.BodyExtractors
 import org.springframework.web.reactive.function.server.*
 import javax.validation.Validator
 import kotlin.coroutines.coroutineContext
-import no.nav.k9.*
-import no.nav.k9.mappe.*
-import no.nav.k9.mappe.MappeService
-import no.nav.k9.søknad.JsonUtils
-import no.nav.k9.søknad.ValideringsFeil
-import no.nav.k9.søknad.felles.*
-import no.nav.k9.søknad.felles.Barn
-import no.nav.k9.søknad.felles.Periode
-import no.nav.k9.søknad.felles.Språk
-import no.nav.k9.søknad.pleiepengerbarn.*
-import no.nav.k9.søknad.pleiepengerbarn.Arbeid
-import no.nav.k9.søknad.pleiepengerbarn.Tilsynsordning
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
-import org.springframework.http.HttpStatus
-import java.time.ZoneId
 
 @Configuration
 internal class PleiepengerSyktBarnRoutes(
@@ -33,7 +24,8 @@ internal class PleiepengerSyktBarnRoutes(
         private val objectMapper: ObjectMapper,
         private val mappeService: MappeService,
         private val pleiepengerSyktBarnSoknadService: PleiepengerSyktBarnSoknadService,
-        private val authenticationHandler: AuthenticationHandler
+        private val authenticationHandler: AuthenticationHandler,
+        private val pleiepengerSyktBarnSoknadConverter: PleiepengerSyktBarnSoknadConverter
 ) {
     private companion object {
         private val logger: Logger = LoggerFactory.getLogger(PleiepengerSyktBarnRoutes::class.java)
@@ -126,7 +118,7 @@ internal class PleiepengerSyktBarnRoutes(
                                 .buildAndAwait()
                         mappeDTO.erKomplett() -> {
                             try {
-                                val soknad = søknadConverter(objectMapper.convertValue(mappe.person[norskIdent]!!.soeknad), norskIdent)
+                                val soknad = pleiepengerSyktBarnSoknadConverter.convert(objectMapper.convertValue(mappe.person[norskIdent]!!.soeknad), norskIdent)
                                 val soknadjson: String = JsonUtils.toString(soknad)
                                 pleiepengerSyktBarnSoknadService.sendSøknad(soknadjson)
                                 mappeService.fjern(
@@ -195,57 +187,4 @@ internal class PleiepengerSyktBarnRoutes(
     }
     private suspend fun ServerRequest.innsending() = body(BodyExtractors.toMono(Innsending::class.java)).awaitFirst()
     private fun ServerRequest.mappeLocation(mappeId: MappeId) = uriBuilder().pathSegment("mappe", mappeId).build()
-
-    private fun søknadConverter(pleiepengerSyktBarnSoknad: PleiepengerSyktBarnSoknad, ident: NorskIdent): PleiepengerBarnSøknad {
-
-        var tilsynsordningBuilder = Tilsynsordning.builder().iTilsynsordning(when (pleiepengerSyktBarnSoknad.tilsynsordning?.iTilsynsordning) {
-            JaNeiVetikke.ja -> TilsynsordningSvar.JA
-            JaNeiVetikke.nei -> TilsynsordningSvar.NEI
-            else -> TilsynsordningSvar.VET_IKKE
-        })
-        pleiepengerSyktBarnSoknad.tilsynsordning?.opphold?.forEach{
-            tilsynsordningBuilder = tilsynsordningBuilder.uke(
-                    TilsynsordningUke.builder()
-                            .periode(periodeConverter(it.periode))
-                            .mandag(it.mandag)
-                            .tirsdag(it.tirsdag)
-                            .onsdag(it.onsdag)
-                            .torsdag(it.torsdag)
-                            .fredag(it.fredag)
-                            .build()
-            )
-        }
-
-        return PleiepengerBarnSøknad.builder()
-                .søknadId(SøknadId.of(pleiepengerSyktBarnSoknad.id))
-                .mottattDato(pleiepengerSyktBarnSoknad.datoMottatt?.atStartOfDay()?.atZone(ZoneId.systemDefault()))
-                .søker(Søker.builder().norskIdentitetsnummer(NorskIdentitetsnummer.of(ident)).build())
-                .barn(Barn.builder()
-                        .norskIdentitetsnummer(NorskIdentitetsnummer.of(pleiepengerSyktBarnSoknad.barn?.norskIdent))
-                        .fødselsdato(pleiepengerSyktBarnSoknad.barn?.foedselsdato)
-                        .build())
-                .språk(Språk.of(pleiepengerSyktBarnSoknad.spraak.toString()))
-                .søknadsperioder(pleiepengerSyktBarnSoknad.perioder?.map{periodeConverter(it) to SøknadsperiodeInfo.builder().build()}?.toMap())
-                .arbeid(Arbeid.builder()
-                        .arbeidstaker(pleiepengerSyktBarnSoknad.arbeid?.arbeidstaker?.map{Arbeidstaker.builder()
-                                .norskIdentitetsnummer(NorskIdentitetsnummer.of(it.norskIdent))
-                                .organisasjonsnummer(Organisasjonsnummer.of(it.organisasjonsnummer))
-                                .perioder(it.skalJobbeProsent?.map{periodeConverter(it.periode) to Arbeidstaker.ArbeidstakerPeriodeInfo.builder().skalJobbeProsent(it.grad?.toBigDecimal()).build()}?.toMap())
-                                .build()})
-                        .frilanser(pleiepengerSyktBarnSoknad.arbeid?.frilanser?.map{Frilanser.builder().periode(periodeConverter(it.periode), Frilanser.FrilanserPeriodeInfo()).build()})
-                        .selvstendigNæringsdrivende(pleiepengerSyktBarnSoknad.arbeid?.selvstendigNaeringsdrivende?.map{SelvstendigNæringsdrivende.builder().periode(periodeConverter(it.periode), SelvstendigNæringsdrivende.SelvstendigNæringsdrivendePeriodeInfo()).build()})
-                        .build())
-                .beredskap(Beredskap.builder()
-                        .perioder(pleiepengerSyktBarnSoknad.beredskap?.map{periodeConverter(it.periode) to Beredskap.BeredskapPeriodeInfo.builder().tilleggsinformasjon(it.tilleggsinformasjon).build()}?.toMap())
-                        .build())
-                .nattevåk(Nattevåk.builder()
-                        .perioder(pleiepengerSyktBarnSoknad.nattevaak?.map{periodeConverter(it.periode) to Nattevåk.NattevåkPeriodeInfo.builder().tilleggsinformasjon(it.tilleggsinformasjon).build()}?.toMap())
-                        .build())
-                .tilsynsordning(tilsynsordningBuilder.build())
-                .build()
-    }
-
-    private fun periodeConverter(periode: no.nav.k9.pleiepengersyktbarn.soknad.Periode?): Periode {
-        return Periode.builder().fraOgMed(periode?.fraOgMed).tilOgMed(periode?.tilOgMed).build();
-    }
 }
